+++ tracpro/__init__.py
@@ -6,7 +6,7 @@ from .celery import app as celery_app  # noqa
 
 
 # NOTE: Version must be updated in docs/conf.py as well.
-VERSION = (1, 4, 3, "final")
+VERSION = (1, 5, 0, "final")
 
 
 def get_version(version):